Implement LOOP, ENDLOOP and BREAK DXBC ops
We also refactor the REP and ENDREP ops to share code with the LOOP AND
ENDLOOP ops. LOOP and ENDLOOP are only partially implemented because
(unlikely) stores to the counter variable are not yet supported. A
future refactoring is expected to support updates to the counter.

VK9's dependence on boost has kept me from cross compiling it in the
time that I could spare, so I have been unable to test this. Chris asked
me to put this in a pull request anyway, so I am doing that.

Signed-off-by: Richard Yao <ryao@gentoo.org>